pyproject.toml
[project]
name = "fkinds-python"
version = "0.1.0"
requires-python = ">=3.14"
description = ""
authors = [
    { name = "fkinds", email = "wkhs.1020@gmail.com" }
]
readme = "README.md"
dependencies = [
    "attrs>=25.4.0",
    "django>=5.2",
    "django-environ>=0.12",
    "djangorestframework>=3.16",
    "fixit>=2.1.0",
    "import-linter>=2.5.2",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "psycopg[binary]>=3.2",
    "pydantic>=2.12.3",
    "pyrefly>=0.53.0",
    "pytest>=8.4.2",
    "ruff>=0.14.3",
    "ty>=0.0.17",
    "uv>=0.10.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-django",
    "pytest-xdist",
    "ruff",
    "mypy",
    "django-stubs",
    "djangorestframework-stubs",
    "pre-commit",
]

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
    "N",      # pep8-naming
    "SIM",    # flake8-simplify
    "T20",    # flake8-print
    "ERA",    # eradicate (commented-out code)
    "PT",     # flake8-pytest-style
    "RET",    # flake8-return
    "ARG",    # flake8-unused-arguments
    "PIE",    # flake8-pie
    "PERF",   # perflint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "RUF012",  # Django Meta / ModelSerializer で ClassVar 不要
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
known-first-party = ["config", "books"]
force-single-line = true

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert の使用を許可
    "ARG",     # 未使用引数 (fixture等)
    "T20",     # print の使用を許可
]
"manage.py" = ["T20"]
"**/migrations/*.py" = ["ALL"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

[[tool.mypy.overrides]]
module = "environ"
ignore_missing_imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["src"]

[tool.ty]

[tool.ty.environment]
python-version = "3.14"

[tool.ty.src]
include = ["src"]

[tool.pyrefly]
project_includes = ["src/**/*.py", "tests/**/*.py"]
python_version = "3.14"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "pytest-xdist",
    "ruff",
    "mypy",
    "django-stubs",
    "djangorestframework-stubs",
    "pre-commit",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
